Fix generation problems, endless game problem, add console ui

+++ include/gamerules.h
@@ -11,26 +11,26 @@
 
 class GameRules {
 private:
-    size_t _widthField;
-    size_t _heightField;
-    size_t _numberOfShips;
+    unsigned _widthField;
+    unsigned _heightField;
+    unsigned _numberOfShips;
     bool _isCustomShips;
-    size_t _turnTime;
+    unsigned _turnTime;
     bool _isActiveBonuses;
-    size_t _squareOfShips;
+    unsigned _squareOfShips;
     std::vector<Bonus> _bonuses;
     std::vector<Ship> _ships;
     Parser parser;
 public:
     GameRules() {};
     GameRules(std::string filename);
-    size_t getWidthField() { return _widthField; };
-    size_t getHeightField() { return _heightField; };
-    size_t getNumberOfShips() { return _numberOfShips; };
-    size_t getTurnTime() { return _turnTime; };
+    unsigned getWidthField() { return _widthField; };
+    unsigned getHeightField() { return _heightField; };
+    unsigned getNumberOfShips() { return _numberOfShips; };
+    unsigned getTurnTime() { return _turnTime; };
     bool getFlagBonuses() { return _isActiveBonuses; };
-    size_t getSquareOfShips() { return _squareOfShips; };
-    Ship* getShip(size_t id);
+    unsigned getSquareOfShips() { return _squareOfShips; };
+    Ship* getShip(unsigned id);
 };
