+++ include/standards.h
@@ -4,7 +4,7 @@
 #include <string>
 
 const std::string STANDARD_CONFIG_PATH = "configs/game.cg";
-const size_t STANDARD_ID_START = 20;
+const unsigned STANDARD_ID_START = 20;
 
 // chapter "rules"
 #define STANDARD_FIELD_WIDTH 10
@@ -15,9 +15,9 @@ const size_t STANDARD_ID_START = 20;
 #define STANDARD_BONUSES false
 
 const std::string STANDARD_SHIPS_NAME[] = { "OneCellShip", "TwoCellShip", "ThreeCellShip", "FourCellShip" };
-const size_t STANDARD_SHIPS_WIDTH[] = { 1, 2, 3, 4 };
-const size_t STANDARD_SHIPS_HEIGHT[] = { 1, 1, 1, 1};
-const size_t STANDARD_SHIPS_ATTACK_RADIUS[] = { 1, 1, 1, 1 };
-const size_t STANDARD_SHIPS_NUMBER[] = { 4, 3, 2, 1 };
+const unsigned STANDARD_SHIPS_WIDTH[] = { 1, 2, 3, 4 };
+const unsigned STANDARD_SHIPS_HEIGHT[] = { 1, 1, 1, 1};
+const unsigned STANDARD_SHIPS_ATTACK_RADIUS[] = { 1, 1, 1, 1 };
+const unsigned STANDARD_SHIPS_NUMBER[] = { 4, 3, 2, 1 };
 
 #endif // WARSHIPS_STANDARDS_H
