+++ src/field.cpp
@@ -1,18 +1,18 @@
 #include "field.h"
 #include "standards.h"
 
-Field::Field(size_t width, size_t height) {
+Field::Field(unsigned width, unsigned height) {
     _numberOfHits = 0;
     _width = width;
     _height = height;
-    for (size_t i = 0; i < width; ++i) {
-        for (size_t j = 0; j < height; ++j) {
+    for (unsigned i = 0; i < width; ++i) {
+        for (unsigned j = 0; j < height; ++j) {
             _cells.emplace_back(i, j, 0);
         }
     }
 }
 
-void Field::setID(size_t x, size_t y, size_t id) {
+void Field::setID(unsigned x, unsigned y, unsigned id) {
     _cells[x*_width + y].setID(id);
     if (id >= STANDARD_ID_START + 10) {                     // TODO(keberson): изменить условие, должно быть не + 10
         _numberOfHits++;
@@ -22,12 +22,12 @@ void Field::setID(size_t x, size_t y, size_t id) {
 std::ostream& operator << (std::ostream& out, Field& field) {
     out.width(3);
     out << ' ';
-    for (size_t i = 0; i < field.getWidth(); ++i) {                // TODO(keberson): предусмотреть случай, когда ширина больше, чем 26
+    for (unsigned i = 0; i < field.getWidth(); ++i) {                // TODO(keberson): предусмотреть случай, когда ширина больше, чем 26
         out.width(3);
         out << (char)('a' + i);
     }
 
-    size_t counter = 0;
+    unsigned counter = 0;
     for (auto item: field._cells) {
         if (item.getY() % field.getWidth() == 0) {
             out << std::endl;
